refactor: 패키지 구조 변경
- 패키지간 의존성을 고려하며 패키지 구조 변경
