+++ src/main/java/lotto/controller/ExceptionHandlingStrategy.java
@@ -1,8 +1,7 @@
-package lotto.exception;
+package lotto.controller;
 
 import java.util.function.BiConsumer;
 import lotto.service.LottoGameService;
-import lotto.view.LottoGameView;
 
 public interface ExceptionHandlingStrategy {
     void apply(LottoGameView lottoGameView, LottoGameService lottoGameService,
