+++ src/main/java/lotto/controller/LottoGameView.java
@@ -1,11 +1,11 @@
-package lotto.view;
+package lotto.controller;
 
 import java.util.List;
-import lotto.dto.LottoBonusNumberCreateRequest;
-import lotto.dto.LottoPurchaseRequest;
-import lotto.dto.LottoResponse;
-import lotto.dto.LottoWinningNumberCreateRequest;
-import lotto.dto.LottoWinningStatistics;
+import lotto.controller.dto.LottoBonusNumberCreateRequest;
+import lotto.controller.dto.LottoPurchaseRequest;
+import lotto.controller.dto.LottoResponse;
+import lotto.controller.dto.LottoWinningNumberCreateRequest;
+import lotto.controller.dto.LottoWinningStatistics;
 
 public interface LottoGameView {
     LottoPurchaseRequest inputPurchaseRequest();
