+++ src/main/java/lotto/controller/dto/LottoPurchaseRequest.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.controller.dto;
 
 import lotto.exception.ExceptionMessage;
