+++ src/main/java/lotto/view/LottoGameConsoleView.java
@@ -2,12 +2,13 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
+import lotto.controller.LottoGameView;
+import lotto.controller.dto.LottoBonusNumberCreateRequest;
+import lotto.controller.dto.LottoPurchaseRequest;
+import lotto.controller.dto.LottoResponse;
+import lotto.controller.dto.LottoWinningNumberCreateRequest;
+import lotto.controller.dto.LottoWinningStatistics;
 import lotto.domain.LottoPrize;
-import lotto.dto.LottoBonusNumberCreateRequest;
-import lotto.dto.LottoPurchaseRequest;
-import lotto.dto.LottoResponse;
-import lotto.dto.LottoWinningNumberCreateRequest;
-import lotto.dto.LottoWinningStatistics;
 
 public class LottoGameConsoleView implements LottoGameView {
